+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Make completion work after pressing `Ctrl + Space` more than once excluding the first time
 - Add completion in Smart Completion by pressing `Ctrl + Shift + Space`
+- Show tool window warnings when translation files in an i18n folder have mismatched key sets
 
 ## [0.0.2]